Move the mesh expansion node agent into the pkg/nodeagent.
